Finally replaced "portablemc.py".
